Created error message component
