+++ src/App.js
@@ -6,6 +6,9 @@ import { PrivateRoute } from "./utils";
 import ExampleComponent from "./components/ExampleComponent";
 import SignUp from "./components/SignUp/";
 import Login from "./components/Login/";
+import Message from './components/Message';
+
+import './App.scss';
 
 const App = () => {
 
@@ -14,7 +17,7 @@ const App = () => {
   }
 
   return (
-    <div>
+    <div className="container">
      <div>
         Welcome To Trip Split
       </div>
@@ -35,6 +38,7 @@ const App = () => {
           <button onClick={logOut}>Log out</button>
         </li>
       </ul>
+      <Message />
       <PrivateRoute path="/secret" component={ExampleComponent} />
       <Route path="/signup" component={SignUp} />  
       <Route path ="/login" component={Login} />
