+++ src/App.scss
@@ -1,5 +1,5 @@
 @import '~antd/dist/antd.css';
 
 .container {
-  margin-left: 10%;
+  margin:5%;
 }
