+++ src/reducers/index.js
@@ -1,4 +1,4 @@
-import { FETCHING, FETCH_SUCCESS, FETCH_FAILURE, LOGGING_IN, LOGIN_FAILURE, LOGIN_SUCCESS, SIGNING_UP, SIGNUP_FAILURE, SIGNUP_SUCCESS, TOGGLE_PAID, SET_EVENT } from '../actions';
+import { FETCHING, FETCH_SUCCESS, FETCH_FAILURE, ERROR, LOGGING_IN, LOGIN_FAILURE, LOGIN_SUCCESS, SIGNING_UP, SIGNUP_FAILURE, SIGNUP_SUCCESS, TOGGLE_PAID, SET_EVENT } from '../actions';
 import { mockData } from '../utils';
 
 export const reducer = (state = initialState, action) => {
@@ -11,18 +11,20 @@ export const reducer = (state = initialState, action) => {
       localStorage.setItem('token', action.payload);
       return {...state, isFetching: false, fetching_message: "", error: false}
     case LOGIN_FAILURE:
-      return {...state, isFetching: false, fetching_message: "", error: true, error_message: action.payload}
+      return {...state, isFetching: false, fetching_message: "", error: true, error_message: "Failed to Login"}
     case SIGNING_UP:
       return {...state, isFetching: true, fetching_message:"SIGNING UP..."}
     case SIGNUP_SUCCESS:
       return {...state, isFetching: false, fetching_message: "SIGNED UP!", error: false}
     case SIGNUP_FAILURE: 
       return {...state, isFetching: false, fetching_message: "", error: true, error_message:"U FAILED"}
-    case TOGGLE_PAID: {
+    case TOGGLE_PAID:
       return {...state, currentData: {...state.currentData, people: action.payload}}
-    } case SET_EVENT: {
+    case SET_EVENT:
       return {...state, singleTrip: action.payload}
-    } default:
+    case ERROR:
+      return {...state, error: true, error_message: action.payload}
+    default:
       return state;
   }
 }
